requirements.txt
# Core
openai>=1.30.0
pinecone>=4.0.0,<6.0.0
pydantic-settings>=2.2.0
python-dotenv>=1.0.0

# Data
pandas>=2.2.0
numpy>=1.26.0

# Retrieval
# bm25s is the active sparse-retrieval backend (numpy-vectorised, ~5-10x faster
# than the legacy rank-bm25 which is no longer imported anywhere).
bm25s>=0.2.0

# Web frontend
flask>=3.0.0
flask-cors>=4.0.0

# Utilities
tqdm>=4.66.0

# Note: the VectifyAI/PageIndex md-to-tree parser is vendored locally in
# pageindex/_md_parser.py (~120 LOC, MIT license) because the upstream
# repo has no setup.py/pyproject.toml and the PyPI `pageindex` package
# is a different (hosted-API) client. No extra dependencies needed.

# Dev / testing (optional)
pytest>=8.0.0
ipython>=8.0.0
